@@ -3,6 +3,7 @@ package org.foxminded.university.dao;
 import org.foxminded.university.domain.Page;
 import org.foxminded.university.domain.Pageable;
 import org.foxminded.university.entity.Address;
+import org.foxminded.university.entity.Schedule;
 import org.foxminded.university.entity.Teacher;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
@@ -21,6 +22,7 @@ public class TeacherDao extends AbstractDao<Long, Teacher> {
     private static final String UPDATE_TEACHER = "UPDATE teachers set firstName = ?, lastName = ?, birthDate = ?, phoneNumber = ?, email = ?, password = ?, address_id = ?, linkedinUrl = ? WHERE id = ?";
     private static final String DELETE_TEACHER = "DELETE FROM teachers WHERE id = ?";
     private static final String FIND_ALL_QUERY_PAGEABLE = "SELECT * FROM teachers ORDER BY id OFFSET ? ROWS FETCH NEXT ? ROWS ONLY";
+    private static final String FIND_ALL_SCHEDULE_FOR_TEACHER = "SELECT schedule_dates.dates, schedules.startTime, schedules.endTime FROM courses LEFT JOIN schedules on courses.schedule_id = schedules.id LEFT JOIN schedule_dates on schedules.id = schedule_dates.schedules_id WHERE courses.teacher_id = ? and schedule_dates.dates is not null";
     private static final RowMapper<Teacher> teacherMapper = (resultSet, rowNum) ->
             Teacher.builder()
                     .withId(resultSet.getLong("id"))
@@ -33,7 +35,12 @@ public class TeacherDao extends AbstractDao<Long, Teacher> {
                     .withAddress(Address.builder().withId(resultSet.getLong("address_id")).build())
                     .withLinkedinUrl(resultSet.getString("linkedinUrl"))
                     .build();
-
+    private static final RowMapper<Schedule> scheduleMapper = (resultSet, rowNum) ->
+            Schedule.builder()
+                    .withDate(resultSet.getDate("schedule_dates.dates").toLocalDate())
+                    .withStartTime(resultSet.getTime("schedules.startTime"))
+                    .withEndTime(resultSet.getTime("schedules.endTime"))
+                    .build();
     @Autowired
     public TeacherDao(JdbcTemplate jdbcTemplate) {
         super(jdbcTemplate);
@@ -69,9 +76,14 @@ public class TeacherDao extends AbstractDao<Long, Teacher> {
     }
 
     @Override
-    Pageable<Teacher> findAll(Page page) {
+    public Pageable<Teacher> findAll(Page page) {
         final int startItems = page.getItemsPerPage() * page.getPageNumber();
         List<Teacher> teachers = jdbcTemplate.query(FIND_ALL_QUERY_PAGEABLE, teacherMapper, startItems, page.getItemsPerPage());
         return new Pageable<>(teachers, page.getPageNumber(), page.getItemsPerPage());
+    }
+
+    public List<Schedule> getScheduleForTeacher(Long id){
+        List<Schedule> dates = jdbcTemplate.query(FIND_ALL_SCHEDULE_FOR_TEACHER, scheduleMapper, id);
+        return dates;
     }
 }
